requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
ConfigArgParse==1.2.3
Faker==5.0.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.5
greenlet==0.4.17
idna==2.10
importlib-metadata==3.1.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
locust==1.4.1
MarkupSafe==1.1.1
msgpack==1.0.1
packaging==20.8
pluggy==0.13.1
psutil==5.7.3
py==1.10.0
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pywin32==300
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
